Add files via upload
This isn't working code yet. I've commented what still needs to be done. However, I wanted to share the classes that I'm working with - especially the Session.java class. I suggest that is the one that we will pass back and forth to gather information about sessions to answer the user.